requirements.txt
flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Script
Flask-WTF
email_validator
Flask-JWT-Extended
openai 
python-dotenv